// generated/addon/angular/src/app/api-models/cieb-group-filing-detail.model.ts
/* Copyright (c) 2021 . All Rights Reserved. */

export class CiebGroupFilingDetail {
  seqGrpfilingId: number;
  seqGroupId: number;
  seqGrpParentId: number;
  filingType: string;
  situsState: string;
  effectiveDate: Date;
  termDate: Date;
  changeReason: string;
  applyToSubgroup: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}